% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix-gridmet-datetimes.R
\name{days_to_datetime}
\alias{days_to_datetime}
\title{Days to Datetimes}
\usage{
days_to_datetime(days)
}
\arguments{
\item{days}{vector of day strings}
}
\value{
vector of datetimes
}
\description{
Days since 1900-01-01 to datetime because the gridMET historical
data netcdfs for some reason don't include a time dimension,
instead include the time in the name of each layer in days since
1900-01-01, ie., 28854 which is 1979-01-01.
}
