% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix-gridmet-datetimes.R
\name{gridmet_names_to_datetime}
\alias{gridmet_names_to_datetime}
\title{gridMET names layer to datetime}
\usage{
gridmet_names_to_datetime(raster)
}
\arguments{
\item{raster}{SpatRaster to convert to datetimes}
}
\value{
vector of day strings
}
\description{
Convert layer names of raster to vector of datetimes
}
