\name{ncep2dt}
\alias{ncep2dt}
\title{Gather global data}
\usage{
  ncep2dt(ans)
}
\arguments{
  \item{ans}{output from \code{\link{NCEP.gather}}
  function}
}
\value{
  data.table with latitude, longitude, and mean
}
\description{
  converts NCEP.gather output to data table
}
\author{
  David LeBauer
}

