\name{panel.cor}
\alias{panel.cor}
\title{pairs.cor}
\usage{
  panel.cor(x, y, digits = 2, prefix = "", cex.cor)
}
\arguments{
  \item{x}{}

  \item{y}{}

  \item{digits}{}

  \item{prefix}{}

  \item{cex.cor}{}
}
\value{
  plot
}
\description{
  enhanced version of pairs()
}
\author{
  R Development Core Team
  http://addictedtor.free.fr/graphiques/graphcode.php?graph=137
}

